small improvement

+++ src/main/java/com/example/demo/logic/SinIntegralLogic.java
@@ -2,7 +2,7 @@ package com.example.demo.logic;
 
 public class SinIntegralLogic {
     public static double counting(double left, double right) {
-        double h = Math.abs((double) (right - left) / 100);
+        double h = Math.abs((right - left) / 100);
         double ans = (Math.sin(left) + Math.sin(right)) / 2;
         for (double i = left + h; i <= right - h + 0.00000001; i += h) {
             ans += Math.sin(i);
